requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2025.11.12
channels==4.3.2
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.1
git-filter-repo==2.47.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.0
packaging==25.0
pillow==12.1.0
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uv==0.9.21
